+++ source/feature-renames.json
@@ -34,7 +34,6 @@
 	"faster-pr-diff-options": "one-click-diff-options",
 	"hide-useless-comments": "hide-low-quality-comments",
 	"no-useless-split-diff-view": "no-unnecessary-split-diff-view",
-	"unwrap-useless-dropdowns": "unwrap-unnecessary-dropdowns",
 	"tag-changelog-link": "tag-changes-link",
 	"list-pr-for-branch": "list-prs-for-branch",
 	"quick-label-hiding": "quick-label-removal",
