+++ source/features/quick-resolve-conflicts.tsx
